add dateSent to list of date converted strings
Right now the 'dateSent' property isn't getting automatically formatted as a date. This field is returned in the `getSms` response callback.

I'd like to add that this method of converting dates should be reconsidered for something more flexible. I'd propose searching *all* fields for things that look datey, or convert to a valid date when they're `Date.parse()`'d. This would add a little processing overhead, though.