Fix webhook url auto detection
We can't just use `request.originalUrl` as the pathname, because it (and valid webhook URIs) could contain a query string. But pathname's aren't allowed to contain a query string, so url.format will inappropriately escape it if it's present. To fix this, we parse the `request.originalUrl` (which is a relative URL) to get whatever path, query, etc. stuff it contains, and then merge that in with the protocol and host.