deps: bump request version to 2.83.x
Updates the request version to 2.83.x to get rid of the uuid deprecation
warning on install.  Request was updated almost a year ago to use a new
uuid module.  Since twilio depends on an older version of request, the
installation continues to emit a deprecation warning.  This commit fixes
the deprecation warning on install.