Amended error message in event API key is provided to Twilio constructor in place of Auth Token
At present, if an auth token is provided instead of an API key the error 'accountSid must start with AC' is issued. This suggests that it is not possible to use API keys in place of auth tokens with twilio-node, when in fact API keys can be used but require an additional optional paramater to be used.

The present error messages encourages poor practice as it indiciates that API keys cannot be used in place of auth tokens.

This change implements a new error message that explains to developer the requirements in order to use API keys in the event that a API key SID is provided (whcih starts with "SK").