Adds property types for AccessToken and Grants
The docs show things like adding an identity to an access token like this:

const accessToken = new AccessToken(
  ACCOUNT_SID,
  API_KEY_SID,
  API_KEY_SECRET,
);

accessToken.identity = 'example';

And similar for options in grants. But the types don't support this and enforce TS users to set them in the options passed to the constructor.

This adds sthe types for the properties for AccessTokens and each *Grant.