fix: webhook validation with array parameter sorting (#723)
Occasionally, sorting of an array parameter will put those parameters in a different order than when the signature was generated on the server. Establishing a sort order for both keys and values of array parameters ensures the same signature is generated when the array values appear in a different order.