fix: update SMS messaging parse error (#855)
Fixes https://github.com/twilio/twilio-node/issues/715

If the messageBody is malformed then there will be a json parse error since the only check to allow parse is whether the body be a string or not. If we wrap the parse in a try catch, we can code a gentler response.