Updated Version each function for better error handling
Added logic to Version each function that catches any errors thrown in the user provided callback and calls the user provided done function with the error. If user does not provide a done function, each will return a promise that will reject with any error that occurs. Added unit tests for the Version each function.