feat: move to built-in URL class
We currently use `url-parse` to deal with URL construction. Given our
node engine constraint of `>=14`, we know `URL` is available in all
supported node versions (introduced in `10.x` if i remember correctly).

So we should be able to drop this dependency and use the built-in parser
instead.

Note that one test here changed hash because url-parse incorrectly did
not encode the single quotes in the test URL. With the standard `URL`
parsing, it is now encoded.